+++ src/DotnetCheckUpdates/Core/NuGetUtils/NuGetPackageSourceProvider.cs
@@ -2,12 +2,15 @@
 // Distributed under the MIT License.
 // https://github.com/vipentti/dotnet-check-updates/blob/main/LICENSE.md
 
+using Microsoft.Extensions.Logging;
 using NuGet.Configuration;
 
 namespace DotnetCheckUpdates.Core.NuGetUtils;
 
-internal class NuGetPackageSourceProvider(IEnumerable<PackageSource>? sources = default)
-    : INuGetPackageSourceProvider
+internal partial class NuGetPackageSourceProvider(
+    ILogger? logger,
+    IEnumerable<PackageSource>? sources = default
+) : INuGetPackageSourceProvider
 {
     private static readonly IEnumerable<PackageSource> s_defaultSources =
     [
@@ -19,12 +22,51 @@ internal class NuGetPackageSourceProvider(IEnumerable<PackageSource>? sources = 
 
     public IEnumerable<PackageSource> GetPackageSources()
     {
-        // Default to nuget v3
-        if (_providedSources.IsEmpty)
-        {
-            return s_defaultSources;
-        }
+        var sources = _providedSources.IsEmpty ? s_defaultSources : _providedSources;
 
-        return _providedSources;
+        foreach (var item in sources)
+        {
+            if (logger is not null)
+            {
+                LogEnabledPackageSource(logger, item.Name, item.SourceUri, item.ProtocolVersion);
+            }
+
+            yield return item;
+
+            // https://learn.microsoft.com/en-us/nuget/reference/nuget-config-file#packagesources
+            // Add json sources additionally with v3 unless they already are
+            if (
+                item.ProtocolVersion != 3
+                && item.Source.EndsWith(
+                    CliConstants.JsonExtensionWithDot,
+                    StringComparison.OrdinalIgnoreCase
+                )
+            )
+            {
+                var itemClone = item.Clone();
+                itemClone.ProtocolVersion = 3;
+                if (logger is not null)
+                {
+                    LogEnabledPackageSource(
+                        logger,
+                        itemClone.Name,
+                        itemClone.SourceUri,
+                        itemClone.ProtocolVersion
+                    );
+                }
+                yield return itemClone;
+            }
+        }
     }
+
+    [LoggerMessage(
+        Level = LogLevel.Debug,
+        Message = "Enabled package source {Name} {SourceUri} {ProtocolVersion}"
+    )]
+    private static partial void LogEnabledPackageSource(
+        ILogger logger,
+        string name,
+        Uri sourceUri,
+        int protocolVersion
+    );
 }